first agg funciton short input type error

Created-by: wangmingyue 00630100
Author-id: 50655
MR-id: 12878095
Commit-by: w00630100
Merged-by: liuyu 00809192
E2E-issues: AR20220715494638
Description:
+ [ ] I've added/changed the documentation
+ [x] I've added test for the code

See merge request Kunpeng-Computing-DC/BigData/OmniOperatorJIT!293

+++ core/src/operator/aggregation/aggregator/aggregator_factory.h
@@ -435,7 +435,7 @@ public:
                     std::move(outputTypes), channels, inputRaw, outputPartial, isOverflowAsNull);
             }
             case OMNI_SHORT: {
-                return std::make_unique<FirstAggregator<IntVector, int16_t>>(aggregateType, std::move(inputTypes),
+                return std::make_unique<FirstAggregator<ShortVector, int16_t>>(aggregateType, std::move(inputTypes),
                     std::move(outputTypes), channels, inputRaw, outputPartial, isOverflowAsNull);
             }
             case OMNI_INT:
